requirements.txt
# requirements.txt
pandas>=2.0.0
numpy>=1.26.0
vnstock==3.2.6
scikit-learn>=1.5.0
ta>=0.10.2
pyyaml>=6.0.1
streamlit>=1.38.0
plotly>=5.22.0
optuna>=3.6.0
torch>=2.4.0
tenacity>=8.0.0
joblib>=1.3.0
matplotlib>=3.7.0